+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking
 
 - `--kube-exclude-type` short flag changed from `-a` to `-t`, `--kube-include-annotation` is `-a` short flag.
+- `--mode` flag renamed to `--provider`.
 
 ### Added
 
@@ -18,10 +19,10 @@
 - Confirmation prompt when running `kahoy apply` without diff or dry-run modes enabled.
 - Optional `--auto-approve` to disable the confirmation prompt.
 - Optional `--create-namespace` on regular and diff modes that will create missing namespaces of applied resources.
+- Kubernetes provider.
 
 ### Changed
 
-- Deprecate `--git-diff-filter` flag in favor of `--include-changes`.
 - `--kube-exclude-type` short flag changed from `-a` to `-t`, `--kube-include-annotation` is `-a` short flag.
 - On Diff, deleted resources now show the real fields and resource the server will delete (before we didn't check the server state).
 - Fix YAML failing on load when YAML file was multiresource and had files only with comments.
@@ -34,6 +35,7 @@
 ### Removed
 
 - Git filtering in favor of generic filtering based on Kubernetes resource diff.
+- `--git-diff-filter` flag in favor of `--include-changes`.
 
 ## [v1.0.0] - 2020-08-31
 
@@ -43,8 +45,8 @@
 - File based filtering (include exclude).
 - Kubernetes type filtering (exclude).
 - Add States repositories (old and new).
-- Paths mode (load from fs).
-- Git mode (load form Git repository).
+- Paths provider (load from fs).
+- Git provider (load form Git repository).
 - Git filtering based on `git diff`.
 - Git states based on previous commit or `git merge-base`.
 - Grouping of resources.